+++ lib/otzaria_icons.dart
@@ -1,3 +1,3 @@
-library otzaria_icons;
+library;
 
 export 'src/generated/otzaria_icons_data.dart';